RBAC: Move resolvers to root
